[TieredStorage] Refactor file_size() code path (#400)
#### Problem
TieredStorage::file_size() essentially supports AccountsFile::len(),
but its API is inconsistent with AccountsFile's.

#### Summary of Changes
Refactor TieredStorage::file_size() to ::len() and share the same API
as AccountsFile's.

#### Test Plan
Build
Existing unit-tests.

+++ accounts-db/src/tiered_storage/readable.rs
@@ -31,7 +31,7 @@ impl TieredStorageReader {
     }
 
     /// Returns the size of the underlying storage.
-    pub fn len(&self) -> u64 {
+    pub fn len(&self) -> usize {
         match self {
             Self::Hot(hot) => hot.len(),
         }
